Refactor task action modals
